requirements.txt
python-docx==1.1.0
colorama==0.4.6
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# ArticleWriter
firecrawl-py
json5
openai